+++ __tests__/api/chatApi.test.ts
@@ -1,0 +1,7 @@
+// TODO
+import { POST } from '@/app/api/chat/route';
+describe('chatApi', () => {
+  it('Should be true', () => {
+    expect(true).toBe(true);
+  });
+});
