+++ __tests__/components/Chat..test.tsx
@@ -1,0 +1,10 @@
+// TODO
+import Chat from "@/app/components/Chat";
+import '@testing-library/jest-dom';
+import { render } from '@testing-library/react';
+
+describe('Chat', () => {
+  it('Should be true', () => {
+    expect(true).toBe(true);
+  })
+});
